Split MIP+IUT code into CLI part in examples and rest in library.
